Use byte array instead of int array.
git-svn-id: file:///home/tomatsu/sandbox/tmp/squawk-mirror/trunk@703 5e201862-2652-49f7-a74a-6bc36218013a

+++ cldc/src/java/lang/FloatingDecimal.java
@@ -1804,7 +1804,7 @@ class FloatingDecimal{
     };
 
     // approximately ceil( log2( long5pow[i] ) )
-    private static final int n5bits[] = {
+    private static final byte n5bits[] = {
 	0,
 	3,
 	5,
